requirements.txt
requests
openai
python-dateutil
packaging
airtable-python-wrapper
fastapi
dotenv
